refactor: MembeRepository 추상화
